# Performance enhancement for string parser