Finds the size of the string and checks whether it needs to be escaped in a single pass, rather than two as before.