@bos 2015-07-21
Looks interesting.

I need a few changes from you, please:
* Drop the `.swp` file
* Squash into a single commit
* Show me what the performance impact is in the commit comment