@bos 2015-07-22
Never mind, I've done the rebasing, squashing, and fixing-up myself :-)