@SeanRBurton 2015-07-22
Thanks, I did some quick benchmarking and it doesn't come out ahead after
all, but I think I can rescue things.
 On 22 Jul 2015 17:26, "bos" <notifications@github.com> wrote:

> Never mind, I've done the rebasing, squashing, and fixing-up myself :-)
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/bos/aeson/pull/271#issuecomment-123780133>.
>