@bos 2015-07-22
Applied as 84ced1530c362646d42fb1f8ea347703859eedc4, thanks.