@bos 2015-07-22
I asked for benchmark results because I expected that your rework would be slow, and the results reveal this to be the case. Here's what I did to close most of the gap: c24f9d37623b066d450358c4aa7b0535bd62b0b9

Unless I find another trick up my sleeve to get past parity, I'll be reverting both commits.