@SeanRBurton 2015-07-22
One problem is that scan is written with a few performance tweaks which
aren't duplicated for runScanner.
On 22 Jul 2015 18:18, "bos" <notifications@github.com> wrote:

> Closed #271 <https://github.com/bos/aeson/pull/271>.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/bos/aeson/pull/271#event-362524236>.
>