@bos 2015-07-22
I don't think those are all that relevant, but I wrote https://github.com/bos/attoparsec/commit/b1c6694447321db82e5a41d36653fcb8dd8d97c9 just to be safe.

I also have some GHC >= 7.8 unboxing tricks up my sleeve that look very promising. Stay tuned, your idea still has legs :-)