@SeanRBurton 2015-07-22
Also note that you can use a three  constructor enum rather than a pair of
bools.
On 22 Jul 2015 21:00, "bos" <notifications@github.com> wrote:

> Performance numbers with GHC 7.10.1 on my MBP, using CompareWithJSON's
> decode/en/aeson/strict result.
>
> Baseline:
>
> time                 1.671 ms   (1.656 ms .. 1.688 ms)
> allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
>   iters              4252843.762 (4251621.741 .. 4254161.104)
>
> @SeanRBurton <https://github.com/SeanRBurton>:
>
> time                 2.054 ms   (2.011 ms .. 2.102 ms)
> allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
>   iters              7174189.077 (7172735.587 .. 7175725.079)
>
> With e9623d0
> <https://github.com/bos/aeson/commit/e9623d0c61fd0f06e538a5491749e47d0d55a517>
> :
>
> time                 1.450 ms   (1.433 ms .. 1.470 ms)
> allocated:           1.000 R²   (1.000 R² .. 1.000 R²)
>   iters              4332621.071 (4331339.889 .. 4333745.585)
>
> Slightly increased memory allocation, but 10% faster than the original
> baseline. Sean, I believe this validates your approach, it just needed a
> little GHC-specific TLC to close the gap. Thanks for thinking of the
> original patch!
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/bos/aeson/pull/271#issuecomment-123846185>.
>